bot_backend/requirements.txt
fastapi==0.95.2
pydantic<2.0
uvicorn==0.22.0
python-dotenv==1.0.0
pymongo==4.5.0
nltk==3.8.1
utils
requests
transformers==4.25.1
torch==1.13.1
pandas==1.5.3
numpy==1.23.5
scikit-learn==1.2.0